getting cmdline profiler working with visual studio 2019

+++ trunk/profilersettings.cpp
@@ -15,7 +15,7 @@ bool ProfilerSettings::SaveAs(std::wstring fname) {
   fprintf(f, FILETYPEID"\n", CURRENT_VERSION);
   fprintf(f, "exe=%s\n", (const char *)m_executable.c_str());
   fprintf(f, "args=%s\n", (const char *)m_commandLineArgs.c_str());
-  fprintf(f, "ndebugPaths=%d\n", m_debugInfoPaths.size());
+  fprintf(f, "ndebugPaths=%d\n", (int)m_debugInfoPaths.size());
   for (std::list<wxString>::iterator it = m_debugInfoPaths.begin();
        it != m_debugInfoPaths.end(); ++it) {
          fprintf(f, "debugpath=%s\n", (const char *)it->c_str());    
@@ -24,14 +24,14 @@ bool ProfilerSettings::SaveAs(std::wstring fname) {
   fprintf(f, "samplingstartdelay=%d\n", m_samplingStartDelay);
   fprintf(f, "samplingtime=%d\n", m_samplingTime);
   fprintf(f, "currentdirectory=%s\n", (const char *)m_currentDirectory.c_str());
-  fprintf(f, "source substitutions=%d\n", m_sourceFileSubstitutions.size());
+  fprintf(f, "source substitutions=%d\n", (int)m_sourceFileSubstitutions.size());
   for (std::map<wxString, wxString>::iterator it = m_sourceFileSubstitutions.begin(); 
        it != m_sourceFileSubstitutions.end(); ++it) {
     fprintf(f, "orig.file=%s\n", (const char *)it->first.c_str());
     fprintf(f, "used file=%s\n", (const char *)it->second.c_str());
   }
 
-  fprintf(f, "symbol abbreviations=%d\n", m_symbolAbbreviations.size());
+  fprintf(f, "symbol abbreviations=%d\n", (int)m_symbolAbbreviations.size());
   for (std::map<wxString, wxString>::iterator it = m_symbolAbbreviations.begin(); 
        it != m_symbolAbbreviations.end(); ++it) {
     fprintf(f, "partial symbol=%s\n", (const char *)it->first.c_str());
@@ -39,7 +39,7 @@ bool ProfilerSettings::SaveAs(std::wstring fname) {
   }
   fprintf(f, "useSymbolServer=%d\n", (int)m_bConnectToSymServer);
 
-  fprintf(f, "environment variables=%d\n", m_environmentVariables.size());
+  fprintf(f, "environment variables=%d\n", (int)m_environmentVariables.size());
   for (std::map<wxString, wxString>::iterator it = m_environmentVariables.begin(); 
        it != m_environmentVariables.end(); ++it) {
     fprintf(f, "variable=%s\n", (const char *)it->first.c_str());
